+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 3.8.7-dev (unreleased)
 ----------------------
+* Feature: Refactored Document's manage template to acquire available 
+  - conversions and qa_scripts with ajax calls [olimpiurob refs #89320]
 * Feature: Enabled the stuck envelopes utility for BDR deployments [robaaoli refs #88393]
 * Bug fix: Bullet proof the checking for activation_log value in stuck envelopes [robaaoli refs #88393]
 * Bug fix: Fixed condition test in EnvelopeInstance's traceActivity method